Remove visibleChannelIds from BisqEasyOfferbookChannelStore and related classes. Remove join/leave from ChatChannelService and move it to PrivateChatChannelService.
