+++ chat/src/main/java/bisq/chat/ChatChannelService.java
@@ -96,12 +96,6 @@ public abstract class ChatChannelService<M extends ChatMessage, C extends ChatCh
 
     public abstract ObservableArray<C> getChannels();
 
-    public void leaveChannel(String id) {
-        findChannel(id).ifPresent(this::leaveChannel);
-    }
-
-    public abstract void leaveChannel(C channel);
-
     protected void doRemoveExpiredMessages(C channel) {
         Set<M> toRemove = channel.getChatMessages().stream()
                 .filter(ChatMessage::isExpired)
