+++ chat/src/main/proto/chat.proto
@@ -146,7 +146,6 @@ message BisqEasyOfferbookChannel {
 }
 message BisqEasyOfferbookChannelStore {
   repeated ChatChannel channels = 1;
-  repeated string visibleChannelIds = 2;
 }
 
 // ChatChannelSelection
